+++ apps/desktop/src/bindings.ts
@@ -91,6 +91,14 @@ async removeWallet(password: string) : Promise<Result<null, ApiError>> {
     else return { status: "error", error: e  as any };
 }
 },
+async resetLocalWallet() : Promise<Result<null, ApiError>> {
+    try {
+    return { status: "ok", data: await TAURI_INVOKE("reset_local_wallet") };
+} catch (e) {
+    if(e instanceof Error) throw e;
+    else return { status: "error", error: e  as any };
+}
+},
 async changeWalletPassword(oldPassword: string, newPassword: string) : Promise<Result<null, ApiError>> {
     try {
     return { status: "ok", data: await TAURI_INVOKE("change_wallet_password", { oldPassword, newPassword }) };